PyTorch/contrib/cv/classification/Centroids-reid/requirements.txt
einops
mlflow
opencv-python
pytorch-lightning==1.1.4
# -f https://download.pytorch.org/whl/torch_stable.html
# torch==1.7.1+cu101
# -f  https://download.pytorch.org/whl/torch_stable.html
torchvision==0.6.0
tqdm
yacs
